WIP: Support Unipi Extensions
According to documentation, Unipi extensions should be handled almost
same way as PLC. One difference is, that we are accessing it through
serial interface.

It make no sense to me to duplicate all code for Neuron, so I split
Board part from neuron.py and I'm using it from neuron and unipi
extension code. Same with features, we can use neuron features with
small changes to detect if we are on extension or PLC.

Fixes #122